Outro validator inutil sendo apagado

+++ stock/src/main/java/victornext/stock/Services/ProductService.java
@@ -2,49 +2,40 @@ package victornext.stock.Services;
 
 
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import victornext.stock.Controller.Mappers.ProductMapper;
+import victornext.stock.Exceptions.DuplicatedException;
 import victornext.stock.Exceptions.NotFoundException;
-import victornext.stock.Model.EnterprisesModel;
 import victornext.stock.Model.ProductModel;
 import victornext.stock.Repositories.ProductRepository;
-import victornext.stock.Repositories.Specs.EnterprisesSpecs;
-import victornext.stock.Repositories.Specs.ProductSpecs;
-import victornext.stock.validators.ProductValidator;
 
 import java.util.List;
-import java.util.Optional;
 
 
 @Service
 @RequiredArgsConstructor
 public class ProductService {
     private final ProductRepository repository;
-    private final ProductValidator validator;
 
 
 
 
-    public ResponseEntity<ProductModel> create(ProductModel model) {
-        //validation
-        validator.ValidatorALL(model);
-        // save
-        ProductModel savedProduct = repository.save(model);
-        //return http status code
-        return ResponseEntity.status(HttpStatus.CREATED).body(savedProduct);
+    public ProductModel create(ProductModel model) {
+        if (repository.existsByNameIgnoreCase(model.getName())) {
+            throw new DuplicatedException("Product with name '" + model.getName() + "' already exists.");
+        }
+        return repository.save(model);
     }
 
 
     public ResponseEntity<?> update(ProductModel model) {
-        validator.ValidatorALL(model);
+        if (repository.existsByNameIgnoreCase(model.getName())) {
+            throw new DuplicatedException("Product with name '" + model.getName() + "' already exists.");
+        }
         ProductModel model1 = repository.save(model);
 
         return ResponseEntity.status(HttpStatus.OK).body(model1);
@@ -52,14 +43,18 @@ public class ProductService {
 
 
     public ResponseEntity<String> delete(Long id) {
-        validator.validateId(id);
+        if(repository.findById(id) == null) {
+            throw new RuntimeException("Nao encontrado");
+        }
         repository.deleteById(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Product successfully deleted");
     }
 
 
     public ResponseEntity<Object> findById(Long id) {
-        validator.validateId(id);
+        if(repository.findById(id) == null) {
+            throw new RuntimeException("Nao encontrado");
+        }
         repository.findById(id);
 
         return ResponseEntity.status(HttpStatus.OK).body(repository.findById(id));
@@ -67,7 +62,9 @@ public class ProductService {
 
 
     public List<ProductModel> Search(String name, Integer page, Integer pageSize) {
-        validator.validateSearchName(name);
+        if (name == null || name.isEmpty()) {
+            throw new NotFoundException("O nome para pesquisa não pode ser vazio ou nulo.");
+        }
 
         Pageable pageRequest = PageRequest.of(page, pageSize);
 
@@ -79,7 +76,6 @@ public class ProductService {
 
 
     public ResponseEntity<Object> AdditionProduct(Long id, Integer quantity) {
-        validator.validateId(id);
         ProductModel model = repository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Product not found!!"));
 
@@ -92,7 +88,9 @@ public class ProductService {
     }
 
     public ResponseEntity<Object> RemoveProduct(Long id, Integer quantity) {
-        validator.validateId(id);
+        if(repository.findById(id) == null) {
+            throw new RuntimeException("Nao encontrado");
+        }
         ProductModel model = repository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Product not found!!"));
